Leaderboard merged into dev

+++ IzvorniKod/backend/src/app.ts
@@ -47,17 +47,22 @@ class App {
 	private async initializeDatabase() {
 		try {
 			await sequelize.authenticate();
-			await sequelize.sync({ force: false });
 			console.log("Connected to the database");
-			this.dbConnected = true;
-			//await this.insertInitialData();
-			//console.log("Initial data inserted successfully");
+
+			if(!this.dbConnected){ //sync only once
+				await sequelize.sync({ force: false });
+				console.log("All models were synchronized successfully.");
+				this.dbConnected = true;
+				/* await this.insertInitialData();
+				console.log("Initial data inserted successfully"); */
+			}
+			
 		} catch (error) {
 			console.error("Unable to connect to the database:", error);
 		}
 	}
 
-	private async insertInitialData() {
+	private async insertInitialData() { //func 4 inserting initial data
 		const { insertRoutes, insertEvents, insertOrganizers, insertAppUsers } = require("./config/database.insert");
 		await insertAppUsers(sequelize.getQueryInterface());
 		await insertOrganizers(sequelize.getQueryInterface());
